changed pid type into int.

+++ interfaces/ISecureShellServer.h
@@ -59,7 +59,7 @@ namespace Exchange {
         virtual uint32_t Activate_dropbear(const string& inputConfigs) = 0;
         virtual uint32_t Deactivate_dropbear() = 0;
         virtual size_t GetActiveSessionsCount() = 0 ;
-        virtual uint32_t CloseClientSession(const uint64_t pid) = 0;
+        virtual uint32_t CloseClientSession(const int pid) = 0;
         virtual IClientSession::IIterator* ActiveClientsIter() = 0;
 
     };
